Rename custom to kpi Move DIFEntityType, DIFMetricType and DIFHostType to a separate file
