+++ pkg/dataingestionframework/data/dif_types.go
@@ -2,40 +2,8 @@ package data
 
 import (
 	"fmt"
-
 	set "github.com/deckarep/golang-set"
-	"github.com/turbonomic/turbo-go-sdk/pkg/proto"
 )
-
-// USING the common DIF Data
-var DIFEntityType = map[proto.EntityDTO_EntityType]string{
-	proto.EntityDTO_VIRTUAL_MACHINE:       "virtualMachine",
-	proto.EntityDTO_APPLICATION_COMPONENT: "application",
-	proto.EntityDTO_BUSINESS_APPLICATION:  "businessApplication",
-	proto.EntityDTO_BUSINESS_TRANSACTION:  "businessTransaction",
-	proto.EntityDTO_DATABASE_SERVER:       "databaseServer",
-	proto.EntityDTO_SERVICE:               "service",
-}
-
-type DIFHostType string
-
-const (
-	VM        DIFHostType = "virtualMachine"
-	CONTAINER DIFHostType = "container"
-)
-
-var DIFMetricType = map[proto.CommodityDTO_CommodityType]string{
-	proto.CommodityDTO_RESPONSE_TIME:     "responseTime",
-	proto.CommodityDTO_TRANSACTION:       "transaction",
-	proto.CommodityDTO_VCPU:              "cpu",
-	proto.CommodityDTO_VMEM:              "memory",
-	proto.CommodityDTO_THREADS:           "threads",
-	proto.CommodityDTO_HEAP:              "heap",
-	proto.CommodityDTO_COLLECTION_TIME:   "collectionTime",
-	proto.CommodityDTO_DB_MEM:            "dbMem",
-	proto.CommodityDTO_DB_CACHE_HIT_RATE: "dbCacheHitRate",
-	proto.CommodityDTO_CONNECTION:        "connection",
-}
 
 //Data ingestion framework topology entity
 type DIFEntity struct {
@@ -126,13 +94,31 @@ func (e *DIFEntity) HostedOnUID(uid string) *DIFEntity {
 	return e
 }
 
-func (e *DIFEntity) Matching(id string) {
+func (e *DIFEntity) Matching(id string) *DIFEntity {
 	if e.MatchingIdentifiers == nil {
 		e.MatchingIdentifiers = &DIFMatchingIdentifiers{id}
-		return
+		return e
 	}
 	// Overwrite
 	e.MatchingIdentifiers.IPAddress = id
+	return e
+}
+
+func (e *DIFEntity) AddMetric(metricType string, key DIFMetricValKey, value float64) {
+	var meList []*DIFMetricVal
+	meList, found := e.Metrics[metricType]
+	if !found {
+		meList = append(meList, &DIFMetricVal{})
+		e.Metrics[metricType] = meList
+	}
+	if len(meList) < 1 {
+		return
+	}
+	if key == AVERAGE {
+		meList[0].Average = &value
+	} else if key == CAPACITY {
+		meList[0].Capacity = &value
+	}
 }
 
 func (e *DIFEntity) String() string {
