+++ pkg/dataingestionframework/schema/dif-entity-schema.json
@@ -129,7 +129,7 @@
         { "$ref": "#/definitions/_connections"},
         { "$ref": "#/definitions/_heap"},
         { "$ref": "#/definitions/_collectionTime"},
-        { "$ref": "#/definitions/_custom"},
+        { "$ref": "#/definitions/_kpi"},
         { "$ref": "#/definitions/_cpu"},
         { "$ref": "#/definitions/_memory"},
         { "$ref": "#/definitions/_threads"},
@@ -269,11 +269,11 @@
       },
       "required": ["memory"]
     },
-    "_custom" : {
+    "_kpi" : {
       "description": "Specification for the custom metrics",
       "type": "object",
       "properties": {
-        "custom": {
+        "kpi": {
           "type": "array",
           "items": {
             "$ref": "#/definitions/metricValueWithKey"
@@ -282,7 +282,7 @@
           "minItems": 1
         }
       },
-      "required": ["custom"]
+      "required": ["kpi"]
     },
 
     "metricValueWithKey": {
